changed unsigned GL types from int to uint

+++ examples/es2/glfw/main.go
@@ -93,16 +93,16 @@ func initScene() {
 	gl.DepthMask(false)
 
 	program := Program(FragmentShader(fsh), VertexShader(vsh))
-	gl.UseProgram(program)
+	gl.UseProgram(uint(program))
 	uMVP = gl.GetUniformLocation(program, "uMVP")
 	attrPos = gl.GetAttribLocation(program, "pos")
 	attrColor = gl.GetAttribLocation(program, "color")
 
-	gl.EnableVertexAttribArray(attrPos)
-	gl.EnableVertexAttribArray(attrColor)
+	gl.EnableVertexAttribArray(uint(attrPos))
+	gl.EnableVertexAttribArray(uint(attrColor))
 	gl.ClearColor(0.5, 0.5, 0.5, 1.0)
-	gl.VertexAttribPointer(attrPos, 2, gl.FLOAT, false, 0, gl.Void(&POSITION))
-	gl.VertexAttribPointer(attrColor, 4, gl.FLOAT, false, 0, gl.Void(&COLOR))
+	gl.VertexAttribPointer(uint(attrPos), 2, gl.FLOAT, false, 0, gl.Void(&POSITION))
+	gl.VertexAttribPointer(uint(attrColor), 4, gl.FLOAT, false, 0, gl.Void(&COLOR))
 
 }
 func drawScene() {
